refactor: modular architecture
