+++ src/domain/index.ts
@@ -1,1 +1,2 @@
-export * from "./chainContext";
+export * as events from "./events";
+export * as polling from "./polling";
