+++ src/domain/polling/filtering/badOrder.ts
@@ -1,12 +1,14 @@
 import { Order } from "@cowprotocol/contracts";
 import { BigNumber, ethers } from "ethers";
 
+const MINIMUM_VALIDITY_SECONDS = 60;
+
 /**
  * Process an order to determine if it is valid
  * @param order The GPv2.Order data struct to validate
  * @throws Error if the order is invalid
  */
-export function validateOrder(order: Order) {
+export function check(order: Order) {
   // amounts must be non-zero
   if (BigNumber.from(order.sellAmount).isZero()) {
     throw new Error("Order has zero sell amount");
@@ -30,8 +32,11 @@ export function validateOrder(order: Order) {
     throw new Error("Order has identical sell and buy token addresses");
   }
 
-  // Check to make sure that the order has at least 120s of validity
-  if (Math.floor(Date.now() / 1000) + 60 > Number(order.validTo)) {
+  // Check to make sure that the order has at least a specified validity
+  if (
+    Math.floor(Date.now() / 1000) + MINIMUM_VALIDITY_SECONDS >
+    Number(order.validTo)
+  ) {
     throw new Error("Order expires too soon");
   }
 }
