+++ src/services/index.ts
@@ -1,0 +1,7 @@
+import { DBService } from "./storage";
+import { ApiService } from "./api";
+import { ChainContext, SDK_BACKOFF_NUM_OF_ATTEMPTS } from "./chain";
+
+// Exporting the `SDK_BACKOFF_NUM_OF_ATTEMPTS` is a smell that can be eliminated
+// when upstream (`cow-sdk`) allows passing instantiated `OrderBookApi` for `.poll`.
+export { DBService, ApiService, ChainContext, SDK_BACKOFF_NUM_OF_ATTEMPTS };
